db connected to localhost

+++ Angkutmi-Backend/database/migrations/2024_11_18_014543_create_vehicles_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('vehicles', function (Blueprint $table) {
-            $table->integer('vehicle_id')->primary();
+            $table->unsignedBigInteger('vehicle_id')->primary();
             $table->enum('type', ['motor', 'pickup', 'truck']);
             $table->integer('capacity');
             $table->string('plat');
@@ -28,3 +28,8 @@ return new class extends Migration
         Schema::dropIfExists('vehicles');
     }
 };
+
+
+
+
+
